runtime: working-group: Refactor budget handler.
