+++ runtime-modules/common/src/lib.rs
@@ -12,10 +12,11 @@ use codec::{Codec, Decode, Encode};
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
 
+use frame_support::dispatch::DispatchResult;
 use frame_support::traits::LockIdentifier;
 use frame_support::Parameter;
 pub use membership::{ActorId, MemberId, MembershipTypes, StakingAccountValidator};
-use sp_arithmetic::traits::BaseArithmetic;
+use sp_arithmetic::traits::{BaseArithmetic, Saturating};
 use sp_runtime::traits::{MaybeSerialize, Member};
 use sp_std::collections::btree_set::BTreeSet;
 use sp_std::vec::Vec;
@@ -99,4 +100,30 @@ pub fn current_block_time<T: frame_system::Trait + pallet_timestamp::Trait>(
 pub trait AllowedLockCombinationProvider {
     /// Return allowed locks combination set.
     fn get_allowed_lock_combinations() -> BTreeSet<(LockIdentifier, LockIdentifier)>;
+}
+
+/// Provides an interface for the generic budget.
+pub trait BudgetManager<AccountId, Balance: Saturating> {
+    /// Returns the current council balance.
+    fn get_budget() -> Balance;
+
+    /// Set the current budget value.
+    fn set_budget(budget: Balance);
+
+    /// Remove some balance from the council budget and transfer it to the account. Fallible.
+    fn try_withdraw(account_id: &AccountId, amount: Balance) -> DispatchResult;
+
+    /// Remove some balance from the council budget and transfer it to the account. Infallible.
+    /// No side-effects on insufficient council balance.
+    fn withdraw(account_id: &AccountId, amount: Balance) {
+        let _ = Self::try_withdraw(account_id, amount);
+    }
+
+    /// Increase the current budget value up to specified amount.
+    fn increase_budget(amount: Balance) {
+        let current_budget = Self::get_budget();
+        let new_budget = current_budget.saturating_add(amount);
+
+        Self::set_budget(new_budget);
+    }
 }
