+++ runtime-modules/working-group/src/lib.rs
@@ -65,6 +65,7 @@ use frame_support::dispatch::DispatchResult;
 use staking_handler::StakingHandler;
 
 type WeightInfoWorkingGroup<T, I> = <T as Trait<I>>::WeightInfo;
+type Balances<T> = balances::Module<T>;
 
 /// Working group WeightInfo
 /// Note: This was auto generated through the benchmark CLI using the `--weight-trait` flag
@@ -1627,8 +1628,8 @@ impl<T: Trait<I>, I: Instance> common::working_group::WorkingGroupAuthenticator<
     }
 }
 
-impl<T: Trait<I>, I: Instance> common::working_group::WorkingGroupBudgetHandler<T>
-    for Module<T, I>
+impl<T: Trait<I>, I: Instance>
+    common::working_group::WorkingGroupBudgetHandler<T::AccountId, BalanceOf<T>> for Module<T, I>
 {
     fn get_budget() -> BalanceOf<T> {
         Self::budget()
@@ -1636,6 +1637,23 @@ impl<T: Trait<I>, I: Instance> common::working_group::WorkingGroupBudgetHandler<
 
     fn set_budget(new_value: BalanceOf<T>) {
         Self::set_working_group_budget(new_value);
+    }
+
+    fn try_withdraw(account_id: &T::AccountId, amount: BalanceOf<T>) -> DispatchResult {
+        ensure!(
+            Self::get_budget() >= amount,
+            Error::<T, I>::InsufficientBalanceForTransfer
+        );
+
+        let _ = Balances::<T>::deposit_creating(account_id, amount);
+
+        let current_budget = Self::get_budget();
+        let new_budget = current_budget.saturating_sub(amount);
+        <Self as common::working_group::WorkingGroupBudgetHandler<T::AccountId, BalanceOf<T>>>::set_budget(
+            new_budget,
+        );
+
+        Ok(())
     }
 }
